Extract shared generator helpers
